+++ test-modules/d/f.js
@@ -1,0 +1,1 @@
+exports.var_of_f = 'var of f';